docs: update Lil Nouns Center link in README.md
The Lil Nouns Center link in the README.md file has been updated. The previous link was "https://www.lil-nouns.center" and it has been changed to "https://lilnouns.center". This change ensures that the Lil Nouns Center serves as the knowledge center and resource hub for Lil Nouns DAO.